Move Page classes into separate files :truck:
