+++ Src/Notion.Client/Models/Page/DatabaseParent.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class DatabaseParent : Parent
+    {
+        [JsonProperty("database_id")]
+        public string DatabaseId { get; set; }
+    }
+}
