+++ Src/Notion.Client/Models/Page/PageParent.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class PageParent : Parent
+    {
+        [JsonProperty("page_id")]
+        public string PageId { get; set; }
+    }
+}
