+++ Src/Notion.Client/Models/Page/WorkspaceParent.cs
@@ -1,0 +1,6 @@
+﻿namespace Notion.Client
+{
+    public class WorkspaceParent : Parent
+    {
+    }
+}
